+++ src/django/requirements.txt
@@ -1,2 +1,4 @@
 Django
 psycopg2-binary # For PostgreSQL database support
+pandas # For reading excel files
+openpyxl # For working with excel files
