+++ src/mysite/dashboard/forms.py
@@ -1,0 +1,4 @@
+from django import forms
+
+class UploadFileForm(forms.Form):
+    csv_file = forms.FileField()
