+++ src/mysite/dashboard/urls.py
@@ -4,4 +4,5 @@ from . import views
 
 urlpatterns = [
     path('', views.dashboard_view, name='dashboard'),
+    path('upload/', views.upload_file, name='upload_file'),
 ]
