+++ src/mysite/mysite/settings.py
@@ -79,14 +79,6 @@ WSGI_APPLICATION = 'mysite.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql',
-        'NAME': os.getenv("ACCOUNTS_DB_NAME"),
-        'USER': os.getenv("ACCOUNTS_DB_USER"),
-        'PASSWORD': os.getenv("ACCOUNTS_DB_PASSWORD"),
-        'HOST': os.getenv("ACCOUNTS_DB_HOST"),
-        'PORT': os.getenv("ACCOUNTS_DB_PORT"),
-    },
-    'bioeconomy_db': {
-        'ENGINE': 'django.db.backends.postgresql',
         'NAME': os.getenv("BIOECONOMY_DB_NAME"),
         'USER': os.getenv("BIOECONOMY_DB_USER"),
         'PASSWORD': os.getenv("BIOECONOMY_DB_PASSWORD"),
